pyproject.toml
[tool.poetry]
name = "fastapitests"
version = "0.1.0"
description = ""
authors = ["Kalimullin Irek <irektopg@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = "^0.118.0"
sqlalchemy = "^2.0.43"
aiosqlite = "^0.21.0"
email-validator = "^2.3.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
pydantic-settings = "^2.11.0"
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.37.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
